+++ contracts/cosmwasm-vm/cw-common/Cargo.toml
@@ -21,7 +21,8 @@ hex = { version = "0.4.3", default-features = false}
 common = { path="../../../libraries/rust/common" }
 ibc-proto = { version = "0.26.0", default-features = false, features = ["parity-scale-codec", "borsh"] }
 prost = { version = "0.11.8", default-features = false,features=["prost-derive"]}
-
+serde-json-wasm = {version="0.5.0", default-features = false}
+bytes = { version = "1.4.0", default-features = false }
 [dev-dependencies]
 getrandom = {version = "0.2.9", default-features = false, features = ["custom"]}
 test-utils={path="../../../libraries/rust/test-utils"}
