+++ contracts/cosmwasm-vm/cw-common/src/constants.rs
@@ -1,7 +1,6 @@
 pub const ICON_CLIENT_STATE_TYPE_URL: &str = "/icon.lightclient.v1.ClientState";
 pub const ICON_CONSENSUS_STATE_TYPE_URL: &str = "/icon.lightclient.v1.ConsensusState";
-pub const ICON_SIGNED_HEADER_TYPE_URL: &str = "/icon.types.v1.SignedHeader";
-pub const ICON_BTP_HEADER_TYPE_URL: &str = "/icon.types.v1.BtpHeader";
+
 pub const ICON_MERKLE_TYPE_URL: &str = "/icon.types.v1.MerkleNode";
 
 pub const ICON_CLIENT_TYPE: &str = "iconclient";
