+++ contracts/cosmwasm-vm/cw-ibc-core/tests/test_connection.rs
@@ -14,11 +14,11 @@ use cosmwasm_std::SubMsgResult;
 use cosmwasm_std::SystemResult;
 use cosmwasm_std::WasmQuery;
 use cw_common::client_response::{OpenAckResponse, OpenConfirmResponse, OpenTryResponse};
+use cw_common::client_state::ClientState;
+use cw_common::consensus_state::ConsensusState;
 use cw_common::types::{ClientId, ConnectionId};
 use cw_ibc_core::constants::*;
 use cw_ibc_core::context::CwIbcCoreContext;
-use cw_ibc_core::ics02_client::types::ClientState;
-use cw_ibc_core::ics02_client::types::ConsensusState;
 use cw_ibc_core::ics03_connection::event::create_open_ack_event;
 use cw_ibc_core::ics03_connection::event::create_open_confirm_event;
 use cw_ibc_core::ics03_connection::event::create_open_init_event;
