+++ contracts/cosmwasm-vm/cw-icon-light-client/src/light_client.rs
@@ -1,13 +1,12 @@
 use crate::traits::AnyTypes;
 use crate::traits::{ConsensusStateUpdate, IContext, ILightClient};
 use crate::ContractError;
+use common::constants::ICON_SIGNED_HEADER_TYPE_URL;
 use common::icon::icon::lightclient::v1::ClientState;
 use common::icon::icon::lightclient::v1::ConsensusState;
 use common::icon::icon::types::v1::{BtpHeader, MerkleNode, SignedHeader};
 use common::utils::{calculate_root, keccak256};
-use cw_common::constants::{
-    ICON_CLIENT_STATE_TYPE_URL, ICON_CONSENSUS_STATE_TYPE_URL, ICON_SIGNED_HEADER_TYPE_URL,
-};
+use cw_common::constants::{ICON_CLIENT_STATE_TYPE_URL, ICON_CONSENSUS_STATE_TYPE_URL};
 use prost::Message;
 
 impl AnyTypes for ClientState {
