+++ libraries/rust/common/src/lib.rs
@@ -1,5 +1,7 @@
 pub mod btp_header;
+pub mod constants;
 pub mod icon;
 pub mod rlp;
+pub mod signed_header;
 pub mod types;
 pub mod utils;
